+++ src/Resque/Job/JobInterface.php
@@ -8,4 +8,14 @@ interface JobInterface
      * @return bool
      */
     public function perform();
+
+    /**
+     * @return void
+     */
+    public function setUp(): void;
+
+    /**
+     * @return void
+     */
+    public function tearDown(): void;
 }