+++ src/Resque/Resque.php
@@ -12,7 +12,7 @@ namespace Resque;
 
 class Resque
 {
-    public const VERSION = '2.6.1';
+    public const VERSION = '2.6.2';
 
     public const DEFAULT_INTERVAL = 5;